Remove titlebar tests and activate unit extended data view tests
