+++ browserTests/constants.js
@@ -1,0 +1,6 @@
+import { Selector } from "testcafe"
+
+
+export const TitleBarTitleSelector = () => Selector('h3[class*="TitleText"]');
+
+export const SearchBarBackButton = () => Selector('#SearchBar .SMBackButton')
